+++ form_designer/models.py
@@ -49,13 +49,23 @@ class Form(models.Model):
     CONFIG_OPTIONS = [
         (
             "save_fs",
-            {"title": _("Save form submission"), "process": create_form_submission},
+            {
+                "title": _("Save form submission"),
+                "description": _(
+                    "Save form submissions in the database"
+                    " so that they may be exported later."
+                ),
+                "process": create_form_submission,
+            },
         ),
         (
             "email",
             {
                 "title": _("Send e-mail"),
-                "form_fields": [
+                "description": _(
+                    "Send the submitted form data to a list of email addresses."
+                ),
+                "form_fields": lambda form: [
                     (
                         "email",
                         forms.CharField(